W
- fixed language to be more accurate (phi -> psi)
- added docstrings for existing functions
- restructured file structure to be compatible with PyPi
- published to PIP!
