+++ CheToolbox.egg-info/PKG-INFO
@@ -1,0 +1,13 @@
+Metadata-Version: 2.1
+Name: CheToolbox
+Version: 1.1
+Summary: A library of commoon chemical engineering calculations
+Author: Quan Phan & Ethan Molnar
+Author-email: Phanqv@mail.uc.edu
+Keywords: python,Chemical Engineering
+Classifier: Intended Audience :: Education
+Classifier: Programming Language :: Python :: 2
+Classifier: Programming Language :: Python :: 3
+Classifier: Operating System :: Microsoft :: Windows
+
+A personal project cataloging common chemical engineering calculations employing numerical methods
