+++ CheToolbox.egg-info/requires.txt
@@ -1,0 +1,2 @@
+numpy
+pandas
