--- CheToolbox/__init__.py
@@ -1,17 +1,0 @@
-"""A Chemical Engineer's Toolbox for Quick Calculations
-
-
-A personal project for quickening my own workflow.
-
-Typical usage example:
-
-import CheToolbox as ChE
-Q = 0.05
-rho = 1000
-mu = 1e-3
-ChE.PressureDrop(Q=Q,rho=rho,mu=mu,D=0.25)
-
-"""
-
-
-from.PressureDrop import PressureDrop
