--- Functions/__main__.py
@@ -1,55 +1,0 @@
-import numpy as np
-import pandas as pd
-
-bootstrings = {
-"bubble_point": r'''
- ________  ___  ___  ________  ________  ___       _______      
-|\   __  \|\  \|\  \|\   __  \|\   __  \|\  \     |\  ___ \     
-\ \  \|\ /\ \  \\\  \ \  \|\ /\ \  \|\ /\ \  \    \ \   __/|    
- \ \   __  \ \  \\\  \ \   __  \ \   __  \ \  \    \ \  \_|/__  
-  \ \  \|\  \ \  \\\  \ \  \|\  \ \  \|\  \ \  \____\ \  \_|\ \ 
-   \ \_______\ \_______\ \_______\ \_______\ \_______\ \_______\
-    \|_______|\|_______|\|_______|\|_______|\|_______|\|_______|
-                                                                
-                                                                
-                                                                
- ________  ________  ___  ________   _________                  
-|\   __  \|\   __  \|\  \|\   ___  \|\___   ___\                
-\ \  \|\  \ \  \|\  \ \  \ \  \\ \  \|___ \  \_|                
- \ \   ____\ \  \\\  \ \  \ \  \\ \  \   \ \  \                 
-  \ \  \___|\ \  \\\  \ \  \ \  \\ \  \   \ \  \                
-   \ \__\    \ \_______\ \__\ \__\\ \__\   \ \__\               
-    \|__|     \|_______|\|__|\|__| \|__|    \|__|               
-''',
-
-"black_mesa": r'''           .-;+$XHHHHHHX$+;-.
-        ,;X@@X%/;=----=:/%X@@X/,
-      =$@@%=.              .=+H@X:
-    -XMX:                      =XMX=
-   /@@:                          =H@+
-  %@X,                            .$@$
- +@X.                               $@%
--@@,                                .@@=
-%@%                                  +@$
-H@:                                  :@H
-H@:         :HHHHHHHHHHHHHHHHHHX,    =@H
-%@%         ;@M@@@@@@@@@@@@@@@@@H-   +@$
-=@@,        :@@@@@@@@@@@@@@@@@@@@@= .@@:
- +@X        :@@@@@@@@@@@@@@@M@@@@@@:%@%
-  $@$,      ;@@@@@@@@@@@@@@@@@M@@@@@@$.
-   +@@HHHHHHH@@@@@@@@@@@@@@@@@@@@@@@+
-    =X@@@@@@@@@@@@@@@@@@@@@@@@@@@@X=
-      :$@@@@@@@@@@@@@@@@@@@M@@@@$:
-        ,;$@@@@@@@@@@@@@@@@@@X/-
-           .-;+$XXHHHHHX$+;-.''',
-
-"phi": r'''
-  _  _  _ 
- | || || |
- | \| |/ |
-  \_   _/ 
-    | |   
-    |_|           
-''',
-
-}
